Add reservations controller to api folder
